requirements.txt
asgiref==3.3.4
astroid==2.5.2
autopep8==1.5.6
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-allauth==0.44.0
django-appconf==1.0.4
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-elasticsearch-dsl==7.2.0
django-filter==2.4.0
django-imagekit==4.0.2
django-parler==2.2
django-parler-rest==2.1
django-rest-auth==0.9.5
django-rest-elasticsearch==0.4.2
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-yasg==1.20.0
elasticsearch==7.13.0
elasticsearch-dsl==7.3.0
Faker==8.6.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
isort==5.8.0
itypes==1.2.0
Jinja2==2.11.3
joblib==1.0.1
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.21.0
oauthlib==3.1.0
packaging==20.9
pandas==1.2.5
pep8==1.7.1
pilkit==2.0
Pillow==8.2.0
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.0.1
pylint==2.7.4
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
scikit-learn==0.24.2
scipy==1.7.0
six==1.15.0
sklearn==0.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
text-unidecode==1.3
threadpoolctl==2.1.0
toml==0.10.2
typed-ast==1.4.2
uritemplate==3.0.1
urllib3==1.26.4
wrapt==1.12.1
